Refactor brand management: - Removed failure state handling in BrandDataTable. - Updated BrandRows to pass brand and cubit as arguments. - Enhanced CreateBrandCubit to manage selected categories and reset fields. - Improved CreateBrandStates to use sealed classes for better state management. - Updated CreateBrandForm to reflect changes in category selection and brand image handling. - Implemented EditBrandCubit for editing brand details with proper state management. - Added BrandFirebaseServices for CRUD operations on brands and categories. - Introduced BrandCategoryModel for managing brand-category relationships. - Updated edit screens to utilize new cubit and model structures.

+++ lib/features/brands/presentation/cubits/brand_cubit.dart
@@ -3,9 +3,7 @@ import 'package:t_store_admin_panel/core/utils/storage/cache_storage_mangement.d
 import 'package:t_store_admin_panel/core/utils/utils/constants/collection_constants.dart';
 import 'package:t_store_admin_panel/data/abstract/cubit/base_data_table_cubit.dart';
 import 'package:t_store_admin_panel/data/abstract/cubit/base_data_table_states.dart';
-import 'package:t_store_admin_panel/data/models/brands/brand_category_model.dart';
 import 'package:t_store_admin_panel/data/models/brands/brand_model.dart';
-import 'package:t_store_admin_panel/data/models/category/category_model.dart';
 import 'package:t_store_admin_panel/data/repositories/brands/brand_repo.dart';
 import 'package:t_store_admin_panel/features/categories/cubits/category/category_cubit.dart';
 
@@ -30,48 +28,52 @@ class BrandCubit extends BaseDataTableCubit<BrandModel> {
 
   @override
   Future<Either<String, void>> deleteItem(BrandModel item) async {
+    if (item.id == null) {
+      return const Left("Item ID is null");
+    }
+
     return await brandRepository.deleteItem(item);
   }
 
   @override
   Future<Either<String, List<BrandModel>>> fetchItems() async {
-    // Fetch All Brands
-    final result = await brandRepository.fetchItems();
-    final fetchedBrands = result.fold(
-      (failure) => <BrandModel>[],
-      (brands) => brands,
-    );
+    // // Fetch All Brands
+    // final result = await brandRepository.fetchItems();
+    // final fetchedBrands = result.fold(
+    //   (failure) => <BrandModel>[],
+    //   (brands) => brands,
+    // );
 
-    // Fetch All BrandCategories
-    final fetchedBrandCategoriesResult =
-        await brandRepository.fetcCategoriesBrand();
-    final fetchedBrandCategories = fetchedBrandCategoriesResult.fold(
-      (failure) => <BrandCategoryModel>[],
-      (list) => list,
-    );
+    // // Fetch All BrandCategories
+    // final fetchedBrandCategoriesResult =
+    //     await brandRepository.fetcCategoriesBrand();
+    // final fetchedBrandCategories = fetchedBrandCategoriesResult.fold(
+    //   (failure) => <BrandCategoryModel>[],
+    //   (list) => list,
+    // );
 
-    // Fetch All Categories
-    final categoriesResult = await categoryCubit.fetchItems();
-    final fetchedCategories = categoriesResult.fold(
-      (failure) => <CategoryModel>[],
-      (list) => list,
-    );
+    // // Fetch All Categories
+    // final categoriesResult = await categoryCubit.fetchItems();
+    // final fetchedCategories = categoriesResult.fold(
+    //   (failure) => <CategoryModel>[],
+    //   (list) => list,
+    // );
 
-    // Set brand categories for each brand
-    for (var brand in fetchedBrands) {
-      final categoryIds =
-          fetchedBrandCategories
-              .where((e) => e.brandId == brand.id)
-              .map((e) => e.categoryId)
-              .toList();
+    // // Set brand categories for each brand
+    // for (var brand in fetchedBrands) {
+    //   final categoryIds =
+    //       fetchedBrandCategories
+    //           .where((e) => e.brandId == brand.id)
+    //           .map((e) => e.categoryId)
+    //           .toList();
 
-      brand.brandCategories =
-          fetchedCategories
-              .where((category) => categoryIds.contains(category.id))
-              .toList();
-    }
+    //   brand.brandCategories =
+    //       fetchedCategories
+    //           .where((category) => categoryIds.contains(category.id))
+    //           .toList();
+    // }
 
-    return result;
+    return await brandRepository.fetchItems();
   }
 
   void sortByColumn(int columnIndex, bool ascending) {
